Ocean Level --> Ocean Value
This is a more precise explanation and I think a better wording. This is really a number to flag a cell as part of the ocean, rather than a sea level, which implies that all cells with lower elevations are also part of the ocean: They would not be, in this scheme. (Such work can be done in preprocessing before computing the DH.)